Rename helpers modules for consistency
