+++ bot.py
@@ -9,15 +9,15 @@ from flask import request, make_response
 
 from slackeventsapi import SlackEventAdapter
 
-from slack_helpers import get_next_change_number
+from helpers.helpers_slack import get_next_change_number
 import settings
 from view_create_change import show_view_create_change
 from view_edit_change import show_view_edit_change
-from slack_helpers import get_slack_client, get_user_list, does_channel_exist
+from helpers.helpers_slack import get_slack_client, get_user_list, does_channel_exist
 from feature_jira import create_jira_release
 from feature_trello import create_trello_release
 from feature_release_notes import post_release_notes, update_release_notes
-from helpers_redis import (
+from helpers.helpers_redis import (
     request_previously_responded,
     request_processed,
     redis_q,
